refactor: split note.service.ts into types, validation, and service modules
Improve code organization by separating concerns:
- types.ts: Type definitions (CreateNoteInput, UpdateNoteInput, etc.)
- validation.ts: Pure validation functions and business logic
- note.service.ts: Handle implementation only
- index.ts: Barrel file for public API exports
