+++ backend/src/services/note/note.service.ts
@@ -11,164 +11,22 @@
  * 5. 関数合成: andThen によるパイプライン処理
  */
 
-import { ok, err, ResultAsync, Result, okAsync, errAsync } from 'neverthrow';
-import { type Note, type NewNote, type NoteWithReplyCount, type Database } from '../../db';
-import { MAX_NOTE_LENGTH } from '@thread-note/shared/constants';
-import { generateId } from '../../utils/id-generator';
-import { extractMentions, getMentionPositions } from '../../utils/mention-parser';
-import {
-  type NoteError,
-  contentTooLongError,
-  contentEmptyError,
-  parentNoteNotFoundError,
-  depthLimitExceededError,
-  circularReferenceError,
-  noteNotFoundError,
-} from '../../errors/domain-errors';
+import { ResultAsync, okAsync, errAsync } from 'neverthrow';
+import { type Note, type NewNote, type Database } from '../../db';
+import { type NoteError, noteNotFoundError } from '../../errors/domain-errors';
 import { createNoteRepository } from '../../repositories/note.repository';
 import { createMentionRepository } from '../../repositories/mention.repository';
+import { generateId } from '../../utils/id-generator';
+import { getMentionPositions } from '../../utils/mention-parser';
 
-// ==========================================
-// Handle Interface (Public API)
-// ==========================================
-
-/** ノート作成のための入力データ */
-export interface CreateNoteInput {
-  readonly content: string;
-  readonly parentId?: string;
-}
-
-/** ノート更新のための入力データ */
-export interface UpdateNoteInput {
-  readonly content: string;
-}
-
-/** ページネーション結果 */
-export interface PaginatedNotes {
-  readonly notes: NoteWithReplyCount[];
-  readonly total: number;
-  readonly hasMore: boolean;
-}
-
-/**
- * NoteService Handle
- *
- * 公開APIを定義するインターフェース。
- * 使用側はこのインターフェースのみを知っていれば良い。
- * テスト時はこのインターフェースをモックする。
- */
-export interface NoteServiceHandle {
-  /** ノートを作成 */
-  readonly createNote: (input: CreateNoteInput) => ResultAsync<Note, NoteError>;
-  /** IDでノートを取得 */
-  readonly getNoteById: (id: string) => ResultAsync<Note, NoteError>;
-  /** ルートノートを取得 */
-  readonly getRootNotes: (
-    limit?: number,
-    offset?: number
-  ) => ResultAsync<PaginatedNotes, NoteError>;
-  /** ノートを更新 */
-  readonly updateNote: (id: string, input: UpdateNoteInput) => ResultAsync<Note, NoteError>;
-  /** ノートを削除（カスケード） */
-  readonly deleteNote: (id: string) => ResultAsync<void, NoteError>;
-}
-
-// ==========================================
-// Internal Helpers (Hidden from outside)
-// ==========================================
-
-/** Result を ResultAsync に変換 */
-const fromResult = <T, E>(result: Result<T, E>): ResultAsync<T, E> =>
-  result.isOk() ? okAsync(result.value) : errAsync(result.error);
-
-/** バリデーション済みノートデータ */
-interface ValidatedNoteData {
-  readonly id: string;
-  readonly content: string;
-  readonly parentId?: string;
-  readonly depth: number;
-  readonly mentionIds: string[];
-}
-
-// ==========================================
-// Pure Validation Functions
-// ==========================================
-
-/**
- * コンテンツの長さを検証
- */
-export const validateContentLength = (content: string): Result<string, NoteError> => {
-  if (content.length === 0) {
-    return err(contentEmptyError());
-  }
-  if (content.length > MAX_NOTE_LENGTH) {
-    return err(contentTooLongError(MAX_NOTE_LENGTH, content.length));
-  }
-  return ok(content);
-};
-
-/**
- * メンションIDを抽出
- */
-export const extractMentionIds = (content: string): string[] => extractMentions(content);
-
-// ==========================================
-// Repository Re-exports
-// ==========================================
-
-export { createNoteRepository } from '../../repositories/note.repository';
-export { createMentionRepository } from '../../repositories/mention.repository';
-
-// ==========================================
-// Pure Business Logic
-// ==========================================
-
-/**
- * 循環参照を検出 (DFS)
- */
-export const detectCircularReference = (
-  graph: Map<string, string[]>,
-  fromNoteId: string,
-  toNoteIds: string[]
-): Result<void, NoteError> => {
-  const tempGraph = new Map(graph);
-  const existingMentions = tempGraph.get(fromNoteId) || [];
-  tempGraph.set(fromNoteId, [...existingMentions, ...toNoteIds]);
-
-  const visited = new Set<string>();
-  const stack = new Set<string>();
-
-  const hasCycle = (nodeId: string): boolean => {
-    if (stack.has(nodeId)) return true;
-    if (visited.has(nodeId)) return false;
-
-    visited.add(nodeId);
-    stack.add(nodeId);
-
-    const neighbors = tempGraph.get(nodeId) || [];
-    for (const neighbor of neighbors) {
-      if (hasCycle(neighbor)) return true;
-    }
-
-    stack.delete(nodeId);
-    return false;
-  };
-
-  return hasCycle(fromNoteId) ? err(circularReferenceError(fromNoteId, toNoteIds)) : ok(undefined);
-};
-
-/**
- * 親ノートから深度を計算
- */
-const calculateDepthFromParent = (
-  parent: Note | undefined,
-  parentId?: string
-): Result<number, NoteError> => {
-  if (!parentId) return ok(0);
-  if (!parent) return err(parentNoteNotFoundError(parentId));
-  if (parent.depth >= 1) return err(depthLimitExceededError(1));
-  return ok(1);
-};
+import type { CreateNoteInput, NoteServiceHandle, ValidatedNoteData } from './types';
+import {
+  fromResult,
+  validateContentLength,
+  extractMentionIds,
+  detectCircularReference,
+  calculateDepthFromParent,
+} from './validation';
 
 // ==========================================
 // Handle Implementation (Factory)
