+++ backend/src/services/note/types.ts
@@ -1,0 +1,75 @@
+/**
+ * Note Service Types
+ *
+ * ノートサービスの型定義。
+ */
+
+import type { ResultAsync } from 'neverthrow';
+import type { Note, NoteWithReplyCount } from '../../db';
+import type { NoteError } from '../../errors/domain-errors';
+
+// ==========================================
+// Input Types
+// ==========================================
+
+/** ノート作成のための入力データ */
+export interface CreateNoteInput {
+  readonly content: string;
+  readonly parentId?: string;
+}
+
+/** ノート更新のための入力データ */
+export interface UpdateNoteInput {
+  readonly content: string;
+}
+
+// ==========================================
+// Output Types
+// ==========================================
+
+/** ページネーション結果 */
+export interface PaginatedNotes {
+  readonly notes: NoteWithReplyCount[];
+  readonly total: number;
+  readonly hasMore: boolean;
+}
+
+// ==========================================
+// Internal Types
+// ==========================================
+
+/** バリデーション済みノートデータ */
+export interface ValidatedNoteData {
+  readonly id: string;
+  readonly content: string;
+  readonly parentId?: string;
+  readonly depth: number;
+  readonly mentionIds: string[];
+}
+
+// ==========================================
+// Handle Interface (Public API)
+// ==========================================
+
+/**
+ * NoteService Handle
+ *
+ * 公開APIを定義するインターフェース。
+ * 使用側はこのインターフェースのみを知っていれば良い。
+ * テスト時はこのインターフェースをモックする。
+ */
+export interface NoteServiceHandle {
+  /** ノートを作成 */
+  readonly createNote: (input: CreateNoteInput) => ResultAsync<Note, NoteError>;
+  /** IDでノートを取得 */
+  readonly getNoteById: (id: string) => ResultAsync<Note, NoteError>;
+  /** ルートノートを取得 */
+  readonly getRootNotes: (
+    limit?: number,
+    offset?: number
+  ) => ResultAsync<PaginatedNotes, NoteError>;
+  /** ノートを更新 */
+  readonly updateNote: (id: string, input: UpdateNoteInput) => ResultAsync<Note, NoteError>;
+  /** ノートを削除（カスケード） */
+  readonly deleteNote: (id: string) => ResultAsync<void, NoteError>;
+}
